fix: enable `http-request-details` feature for indexer crates
The feature was present in the old `indexer/testing` Cargo.toml but not
carried over when collapsing into `dango/testing`, causing the
`transactions_stores_httpd_details` test to fail.
