+++ dango/testing/src/lib.rs
@@ -5,7 +5,10 @@ mod crypto;
 mod genesis;
 pub mod httpd;
 mod hyperlane;
+mod indexer;
 pub mod perps;
 mod setup;
 
-pub use {account::*, account_creation::*, crypto::*, genesis::*, hyperlane::*, setup::*};
+pub use {
+    account::*, account_creation::*, crypto::*, genesis::*, hyperlane::*, indexer::*, setup::*,
+};
