+++ dango/testing/tests/graphql/datetime.rs
@@ -1,7 +1,7 @@
 use {
+    dango_testing::{build_app_service, call_graphql_query, create_block},
     graphql_client::GraphQLQuery,
     indexer_graphql_types::{Block, block},
-    indexer_testing::{block::create_block, build_app_service, call_graphql_query},
 };
 
 #[tokio::test(flavor = "multi_thread")]
