+++ dango/testing/tests/graphql/grug.rs
@@ -1,16 +1,16 @@
 use {
     assert_json_diff::assert_json_eq,
     assertor::*,
+    dango_testing::{
+        GraphQLCustomRequest, build_app_service, call_graphql_query, call_graphql_with_headers,
+        call_ws_graphql_stream, create_block, parse_graphql_subscription_response,
+    },
     graphql_client::GraphQLQuery,
     grug_types::{
         BroadcastClientExt, Coins, Denom, GasOption, Inner, Json, JsonSerExt, Message, Query,
         QueryAppConfigRequest, QueryBalanceRequest, ResultExt,
     },
     indexer_graphql_types::{QueryApp, SubscribeQueryApp, query_app, subscribe_query_app},
-    indexer_testing::{
-        GraphQLCustomRequest, block::create_block, build_app_service, call_graphql_query,
-        call_graphql_with_headers, call_ws_graphql_stream, parse_graphql_subscription_response,
-    },
     serde_json::json,
     std::str::FromStr,
     tokio::sync::mpsc,
