+++ dango/testing/tests/graphql/message.rs
@@ -1,14 +1,13 @@
 use {
     assertor::*,
+    dango_testing::{
+        GraphQLCustomRequest, PaginationDirection, build_app_service, call_graphql_query,
+        call_ws_graphql_stream, create_block, create_blocks, messages_query, paginate_messages,
+        parse_graphql_subscription_response,
+    },
     graphql_client::GraphQLQuery,
     grug_types::{BroadcastClientExt, Coins, Denom, GasOption, Message, ResultExt},
     indexer_graphql_types::{Messages, SubscribeMessages, messages, subscribe_messages},
-    indexer_testing::{
-        GraphQLCustomRequest, PaginationDirection,
-        block::{create_block, create_blocks},
-        build_app_service, call_graphql_query, call_ws_graphql_stream, messages_query,
-        paginate_messages, parse_graphql_subscription_response,
-    },
     std::str::FromStr,
     tokio::sync::mpsc,
 };
