+++ dango/testing/tests/httpd/accounts.rs
@@ -6,7 +6,8 @@ use {
     assert_json_diff::assert_json_eq,
     assertor::*,
     dango_testing::{
-        HyperlaneTestSuite, TestOption, add_account_with_existing_user, create_user_and_account,
+        GraphQLCustomRequest, HyperlaneTestSuite, TestOption, add_account_with_existing_user,
+        call_ws_graphql_stream, create_user_and_account, parse_graphql_subscription_response,
         setup_test_with_indexer,
     },
     dango_types::{
@@ -21,9 +22,6 @@ use {
         QueryBalanceRequest, QueryResponse, QueryWasmSmartRequest, ResultExt,
     },
     indexer_graphql_types::{QueryApp, SubscribeAccounts, query_app, subscribe_accounts},
-    indexer_testing::{
-        GraphQLCustomRequest, call_ws_graphql_stream, parse_graphql_subscription_response,
-    },
     std::collections::BTreeSet,
     tokio::{sync::mpsc, time::sleep},
 };
