+++ dango/testing/tests/indexer_infra/main.rs
@@ -1,2 +1,3 @@
+pub mod api;
 pub mod hooked;
 pub mod sql;
