+++ dango/testing/tests/indexer_infra/sql.rs
@@ -1,6 +1,6 @@
 use {
-    crate::sql::block::replier,
     assertor::*,
+    dango_testing::create_hooked_indexer,
     dango_types::config::AppConfig,
     grug_app::{Db, Indexer},
     grug_testing::TestBuilder,
@@ -11,7 +11,6 @@ use {
     grug_vm_rust::ContractBuilder,
     indexer_cache::cache_file::CacheFile,
     indexer_sql::entity,
-    indexer_testing::setup::create_hooked_indexer,
     replier::{ExecuteMsg, ReplyMsg},
     sea_orm::{ColumnTrait, EntityTrait, QueryFilter, QueryOrder},
     std::str::FromStr,
@@ -21,9 +20,9 @@ use {
 async fn index_block() {
     let denom = Denom::from_str("ugrug").unwrap();
 
-    let (hooked_indexer, indexer_context, _) = create_hooked_indexer().await;
+    let (indexer, sql_indexer_context, ..) = create_hooked_indexer().await;
 
-    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(hooked_indexer)
+    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
         .add_account("owner", Coins::new())
         .add_account("sender", Coins::one(denom.clone(), 30_000).unwrap())
         .set_owner("owner")
@@ -50,27 +49,27 @@ async fn index_block() {
 
     // ensure block was saved
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&sql_indexer_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
     assert_that!(block.unwrap().block_height).is_equal_to(1);
 
     let transactions = entity::transactions::Entity::find()
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch transactions");
     assert_that!(transactions).is_not_empty();
 
     let messages = entity::messages::Entity::find()
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch messages");
 
     assert_that!(messages).is_not_empty();
 
     let events = entity::events::Entity::find()
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch events");
 
@@ -89,7 +88,7 @@ async fn index_block() {
 async fn parse_previous_block_after_restart() {
     let denom = Denom::from_str("ugrug").unwrap();
 
-    let (indexer, indexer_context, _) = create_hooked_indexer().await;
+    let (indexer, sql_indexer_context, ..) = create_hooked_indexer().await;
 
     let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
         .set_app_config(&AppConfig::default())
@@ -110,13 +109,6 @@ async fn parse_previous_block_after_restart() {
         .await
         .should_succeed();
 
-    suite
-        .app
-        .indexer
-        .wait_for_finish()
-        .await
-        .expect("Can't wait for indexer to finish");
-
     // Force the runtime to shutdown or when reusing this `start` would fail
     suite
         .app
@@ -125,16 +117,14 @@ async fn parse_previous_block_after_restart() {
         .await
         .expect("Can't shutdown indexer");
 
-    tracing::warn!("Shut down indexer");
-
     // 1. Delete database block height 1
-    entity::blocks::Entity::delete_block_and_data(&indexer_context.db, 1)
+    entity::blocks::Entity::delete_block_and_data(&sql_indexer_context.db, 1)
         .await
         .unwrap();
 
     // 1 bis. Verify the block height 1 is deleted
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&sql_indexer_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_none();
@@ -149,11 +139,9 @@ async fn parse_previous_block_after_restart() {
         .await
         .expect("Can't start indexer");
 
-    tracing::warn!("Start indexer");
-
     // 4. Verify the block height 1 is indexed
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&sql_indexer_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -180,7 +168,7 @@ async fn parse_previous_block_after_restart() {
     // 6. Verify the block height 2 is indexed
     let block = entity::blocks::Entity::find()
         .order_by_desc(entity::blocks::Column::BlockHeight)
-        .one(&indexer_context.db)
+        .one(&sql_indexer_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -212,13 +200,6 @@ async fn no_sql_index_error_after_restart() {
         )
         .await
         .should_succeed();
-
-    suite
-        .app
-        .indexer
-        .wait_for_finish()
-        .await
-        .expect("Can't wait for indexer to finish");
 
     // Force the runtime to shutdown or when reusing this `start` would fail
     suite
@@ -258,8 +239,6 @@ async fn no_sql_index_error_after_restart() {
     block_to_index
         .save_to_disk()
         .expect("Can't save block on disk");
-
-    tracing::info!("Starting indexer again");
 
     // 3. Start the indexer
     suite
@@ -309,10 +288,147 @@ async fn no_sql_index_error_after_restart() {
     assert!(!block.app_hash.is_empty());
 }
 
+pub mod replier {
+    use {
+        grug_storage::Set,
+        grug_types::{
+            Coins, Empty, ImmutableCtx, Json, JsonSerExt, Message, MutableCtx, Order, QueryRequest,
+            ReplyOn, Response, StdError, StdResult, SubMessage, SubMsgResult, SudoCtx,
+        },
+        serde::{Deserialize, Serialize},
+    };
+
+    #[derive(Serialize, Deserialize)]
+    pub enum ReplyMsg {
+        Ok(ExecuteMsg),
+        Fail(ExecuteMsg),
+    }
+
+    #[derive(Serialize, Deserialize)]
+    pub enum ExecuteMsg {
+        /// Insert the given string into storage. Should be successful.
+        Ok { deep: String },
+        /// Intentionally fail with the given error message.
+        Fail { err: String },
+        /// Insert the given string into storage; then, call self with the given
+        /// execute message.
+        Perform {
+            deep: String,
+            // Must be boxed due to being a recursive type.
+            next: Box<ExecuteMsg>,
+            reply_on: ReplyOn,
+        },
+    }
+
+    impl ExecuteMsg {
+        pub fn ok<T>(deep: T) -> Self
+        where
+            T: Into<String>,
+        {
+            Self::Ok { deep: deep.into() }
+        }
+
+        pub fn perform<T>(deep: T, next: ExecuteMsg, reply_on: ReplyOn) -> Self
+        where
+            T: Into<String>,
+        {
+            Self::Perform {
+                deep: deep.into(),
+                next: Box::new(next),
+                reply_on,
+            }
+        }
+    }
+
+    #[derive(Serialize, Deserialize)]
+    pub enum QueryMsg {
+        Data {},
+    }
+
+    pub struct QueryDataRequest {}
+
+    impl QueryRequest for QueryDataRequest {
+        type Message = QueryMsg;
+        type Response = Vec<String>;
+    }
+
+    impl From<QueryDataRequest> for QueryMsg {
+        fn from(_req: QueryDataRequest) -> Self {
+            Self::Data {}
+        }
+    }
+
+    pub const DEPTHS: Set<String> = Set::new("s");
+
+    pub fn instantiate(_ctx: MutableCtx, _msg: Empty) -> StdResult<Response> {
+        Ok(Response::new())
+    }
+
+    pub fn execute(ctx: MutableCtx, msg: ExecuteMsg) -> StdResult<Response> {
+        match msg {
+            ExecuteMsg::Fail { err } => {
+                // We don't have a generic error as in CosmWasm, so use host
+                // error to mock it.
+                Err(StdError::host(err))
+            },
+            ExecuteMsg::Ok { deep } => {
+                DEPTHS.insert(ctx.storage, deep)?;
+
+                Ok(Response::new())
+            },
+            ExecuteMsg::Perform {
+                deep,
+                next,
+                reply_on,
+            } => {
+                DEPTHS.insert(ctx.storage, deep)?;
+
+                Ok(Response::new().add_submessage(SubMessage {
+                    msg: Message::execute(ctx.contract, &*next, Coins::new())?,
+                    reply_on,
+                }))
+            },
+        }
+    }
+
+    pub fn query(ctx: ImmutableCtx, msg: QueryMsg) -> StdResult<Json> {
+        match msg {
+            QueryMsg::Data {} => {
+                let res = DEPTHS
+                    .range(ctx.storage, None, None, Order::Ascending)
+                    .collect::<StdResult<Vec<_>>>()?;
+                res.to_json_value()
+            },
+        }
+    }
+
+    pub fn reply(ctx: SudoCtx, msg: ReplyMsg, res: SubMsgResult) -> StdResult<Response> {
+        let msg = match (res, msg) {
+            (Result::Err(_), ReplyMsg::Fail(execute_msg))
+            | (Result::Ok(_), ReplyMsg::Ok(execute_msg)) => execute_msg,
+            _ => panic!("invalid reply"),
+        };
+
+        execute(
+            MutableCtx {
+                storage: ctx.storage,
+                api: ctx.api,
+                querier: ctx.querier,
+                chain_id: ctx.chain_id,
+                block: ctx.block,
+                contract: ctx.contract,
+                sender: ctx.contract,
+                funds: Coins::new(),
+            },
+            msg,
+        )
+    }
+}
+
 /// Ensure that flatten events are indexed correctly.
 #[tokio::test(flavor = "multi_thread")]
 async fn index_block_events() {
-    let (indexer, indexer_context, _) = create_hooked_indexer().await;
+    let (indexer, sql_indexer_context, ..) = create_hooked_indexer().await;
 
     let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
         .add_account("owner", Coin::new("usdc", 100_000).unwrap())
@@ -361,27 +477,27 @@ async fn index_block_events() {
 
     // ensure block was saved
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&sql_indexer_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
     assert_that!(block.unwrap().block_height).is_equal_to(1);
 
     let transactions = entity::transactions::Entity::find()
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch transactions");
     assert_that!(transactions).is_not_empty();
 
     let messages = entity::messages::Entity::find()
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch messages");
     assert_that!(messages).is_not_empty();
 
     let events = entity::events::Entity::find()
         .filter(entity::events::Column::BlockHeight.eq(2))
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch events");
     assert_that!(events).is_not_empty();
@@ -395,8 +511,67 @@ async fn index_block_events() {
     // check for parent events
     let events = entity::events::Entity::find()
         .filter(entity::events::Column::ParentId.is_not_null())
-        .all(&indexer_context.db)
+        .all(&sql_indexer_context.db)
         .await
         .expect("Can't fetch events");
     assert_that!(events).is_not_empty();
+}
+
+/// Ensure the indexed blocks are compressed on disk.
+#[tokio::test(flavor = "multi_thread")]
+async fn blocks_on_disk_compressed() {
+    let (indexer, _, cache_context) = create_hooked_indexer().await;
+
+    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
+        .add_account("owner", Coin::new("usdc", 100_000).unwrap())
+        .add_account("sender", Coins::new())
+        .set_owner("owner")
+        .build();
+
+    // Just create a block.
+    let replier_code = ContractBuilder::new(Box::new(replier::instantiate))
+        .with_execute(Box::new(replier::execute))
+        .with_query(Box::new(replier::query))
+        .with_reply(Box::new(replier::reply))
+        .build();
+
+    let _ = suite
+        .upload_and_instantiate(
+            &mut accounts["owner"],
+            replier_code,
+            &Empty {},
+            "salt",
+            Some("label"),
+            None,
+            Coins::default(),
+        )
+        .await
+        .should_succeed()
+        .address;
+
+    // Force the runtime to wait for the async indexer task to finish
+    suite
+        .app
+        .indexer
+        .wait_for_finish()
+        .await
+        .expect("Can't wait for indexer to finish");
+
+    let mut block_path = cache_context.indexer_path.block_path(1);
+
+    block_path.set_extension("borsh.xz");
+
+    assert!(
+        block_path.exists(),
+        "Compressed block_file should exist: {}",
+        block_path.display()
+    );
+
+    block_path.set_extension("");
+
+    assert!(
+        !block_path.exists(),
+        "Decompressed block_file should not exist: {}",
+        block_path.display()
+    );
 }
