--- indexer/testing/tests/indexing/hooked.rs
@@ -1,1 +1,0 @@
-pub mod block;
