--- indexer/testing/tests/indexing/sql.rs
@@ -1,1 +1,0 @@
-pub mod block;
